// js-packages/openk9-admin-new/src/graphql-generated.ts
export type Maybe<T> = T | null;
// ...
export type Analyzer = {
  __typename?: 'Analyzer';
  charFilters?: Maybe<Connection_CharFilter>;
  /** ISO-8601 */
  createDate?: Maybe<Scalars['DateTime']>;
  description?: Maybe<Scalars['String']>;
  id?: Maybe<Scalars['ID']>;
  jsonConfig?: Maybe<Scalars['String']>;
  /** ISO-8601 */
  modifiedDate?: Maybe<Scalars['DateTime']>;
  name?: Maybe<Scalars['String']>;
  tokenFilters?: Maybe<Connection_TokenFilter>;
  tokenizer?: Maybe<Tokenizer>;
  type?: Maybe<Scalars['String']>;
};
// ...
export type Annotator = {
  __typename?: 'Annotator';
  /** ISO-8601 */
  createDate?: Maybe<Scalars['DateTime']>;
  description?: Maybe<Scalars['String']>;
  docTypeField?: Maybe<DocTypeField>;
  docTypeFieldNotInAnnotator?: Maybe<Connection_DocTypeField>;
  extraParams?: Maybe<Scalars['String']>;
  fieldName?: Maybe<Scalars['String']>;
  fuziness?: Maybe<Fuzziness>;
  id?: Maybe<Scalars['ID']>;
  /** ISO-8601 */
  modifiedDate?: Maybe<Scalars['DateTime']>;
  name?: Maybe<Scalars['String']>;
  size?: Maybe<Scalars['Int']>;
  type?: Maybe<AnnotatorType>;
};
// ...
export type Bucket = {
  __typename?: 'Bucket';
  catIndices?: Maybe<Array<Maybe<CatResponse>>>;
  /** ISO-8601 */
  createDate?: Maybe<Scalars['DateTime']>;
  datasources?: Maybe<Connection_Datasource>;
  description?: Maybe<Scalars['String']>;
  docCount?: Maybe<Scalars['BigInteger']>;
  enabled: Scalars['Boolean'];
  id?: Maybe<Scalars['ID']>;
  indexCount?: Maybe<Scalars['BigInteger']>;
  language?: Maybe<Language>;
  languages?: Maybe<Connection_Language>;
  /** ISO-8601 */
  modifiedDate?: Maybe<Scalars['DateTime']>;
  name?: Maybe<Scalars['String']>;
  queryAnalysis?: Maybe<QueryAnalysis>;
  ragConfigurationChat?: Maybe<RagConfiguration>;
  ragConfigurationChatTool?: Maybe<RagConfiguration>;
  ragConfigurationSimpleGenerate?: Maybe<RagConfiguration>;
  refreshOnDate?: Maybe<Scalars['Boolean']>;
  refreshOnQuery?: Maybe<Scalars['Boolean']>;
  refreshOnSuggestionCategory?: Maybe<Scalars['Boolean']>;
  refreshOnTab?: Maybe<Scalars['Boolean']>;
  retrieveType?: Maybe<RetrieveType>;
  searchConfig?: Maybe<SearchConfig>;
  sortings?: Maybe<Connection_Sorting>;
  suggestionCategories?: Maybe<Connection_SuggestionCategory>;
  tabs?: Maybe<Connection_Tab>;
};
// ...
export type CharFilter = {
  __typename?: 'CharFilter';
  /** ISO-8601 */
  createDate?: Maybe<Scalars['DateTime']>;
  description?: Maybe<Scalars['String']>;
  id?: Maybe<Scalars['ID']>;
  jsonConfig?: Maybe<Scalars['String']>;
  /** ISO-8601 */
  modifiedDate?: Maybe<Scalars['DateTime']>;
  name?: Maybe<Scalars['String']>;
  type?: Maybe<Scalars['String']>;
};
// ...
export type Connection_Analyzer = {
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<Edge_Analyzer>>>;
  /** details about this specific page */
  pageInfo?: Maybe<PageInfo>;
};

/** A connection to a list of items. */
export type Connection_Annotator = {
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<Edge_Annotator>>>;
  /** details about this specific page */
  pageInfo?: Maybe<PageInfo>;
};

/** A connection to a list of items. */
export type Connection_Bucket = {
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<Edge_Bucket>>>;
  /** details about this specific page */
  pageInfo?: Maybe<PageInfo>;
};

/** A connection to a list of items. */
export type Connection_CharFilter = {
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<Edge_CharFilter>>>;
  /** details about this specific page */
  pageInfo?: Maybe<PageInfo>;
};

/** A connection to a list of items. */
export type Connection_DataIndex = {
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<Edge_DataIndex>>>;
  /** details about this specific page */
  pageInfo?: Maybe<PageInfo>;
};

/** A connection to a list of items. */
export type Connection_Datasource = {
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<Edge_Datasource>>>;
  /** details about this specific page */
  pageInfo?: Maybe<PageInfo>;
};

/** A connection to a list of items. */
export type Connection_DocType = {
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<Edge_DocType>>>;
  /** details about this specific page */
  pageInfo?: Maybe<PageInfo>;
};

/** A connection to a list of items. */
export type Connection_DocTypeField = {
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<Edge_DocTypeField>>>;
  /** details about this specific page */
  pageInfo?: Maybe<PageInfo>;
};
// ...
export type DataIndex = {
  __typename?: 'DataIndex';
  cat?: Maybe<CatResponse>;
  chunkType?: Maybe<ChunkType>;
  chunkWindowSize?: Maybe<Scalars['Int']>;
  /** ISO-8601 */
  createDate?: Maybe<Scalars['DateTime']>;
  datasource?: Maybe<Datasource>;
  description?: Maybe<Scalars['String']>;
  docCount?: Maybe<Scalars['BigInteger']>;
  docTypes?: Maybe<Connection_DocType>;
  embeddingDocTypeField?: Maybe<DocTypeField>;
  embeddingJsonConfig?: Maybe<Scalars['String']>;
  id?: Maybe<Scalars['ID']>;
  knnIndex?: Maybe<Scalars['Boolean']>;
  mappings?: Maybe<Scalars['String']>;
  /** ISO-8601 */
  modifiedDate?: Maybe<Scalars['DateTime']>;
  name?: Maybe<Scalars['String']>;
  settings?: Maybe<Scalars['String']>;
};
// ...
export type DefaultConnection_Annotator = Connection_Annotator & {
  __typename?: 'DefaultConnection_Annotator';
  edges?: Maybe<Array<Maybe<Edge_Annotator>>>;
  pageInfo?: Maybe<PageInfo>;
};
// ...
export type DefaultConnection_Rule = Connection_Rule & {
  __typename?: 'DefaultConnection_Rule';
  edges?: Maybe<Array<Maybe<Edge_Rule>>>;
  pageInfo?: Maybe<PageInfo>;
};
// ...
export type DefaultEdge_Annotator = Edge_Annotator & {
  __typename?: 'DefaultEdge_Annotator';
  cursor?: Maybe<Scalars['String']>;
  node?: Maybe<Annotator>;
};
// ...
export type DefaultEdge_Rule = Edge_Rule & {
  __typename?: 'DefaultEdge_Rule';
  cursor?: Maybe<Scalars['String']>;
  node?: Maybe<Rule>;
};
// ...
export type DocType = {
  __typename?: 'DocType';
  /** ISO-8601 */
  createDate?: Maybe<Scalars['DateTime']>;
  description?: Maybe<Scalars['String']>;
  docTypeFields?: Maybe<Connection_DocTypeField>;
  docTypeTemplate?: Maybe<DocTypeTemplate>;
  id?: Maybe<Scalars['ID']>;
  /** ISO-8601 */
  modifiedDate?: Maybe<Scalars['DateTime']>;
  name?: Maybe<Scalars['String']>;
};
// ...
export type DocTypeTemplate = {
  __typename?: 'DocTypeTemplate';
  compiled?: Maybe<Scalars['String']>;
  /** ISO-8601 */
  createDate?: Maybe<Scalars['DateTime']>;
  description?: Maybe<Scalars['String']>;
  id?: Maybe<Scalars['ID']>;
  /** ISO-8601 */
  modifiedDate?: Maybe<Scalars['DateTime']>;
  name?: Maybe<Scalars['String']>;
  source?: Maybe<Scalars['String']>;
  templateType?: Maybe<TemplateType>;
};
// ...
export enum Fuzziness {
  Auto = 'AUTO',
  One = 'ONE',
  Two = 'TWO',
  Zero = 'ZERO'
}
// ...
export type Query = {
  __typename?: 'Query';
  analyzer?: Maybe<Analyzer>;
  analyzers?: Maybe<Connection_Analyzer>;
  annotator?: Maybe<Annotator>;
  annotators?: Maybe<Connection_Annotator>;
  bucket?: Maybe<Bucket>;
  buckets?: Maybe<Connection_Bucket>;
  charFilter?: Maybe<CharFilter>;
  charFilters?: Maybe<Connection_CharFilter>;
  dataIndex?: Maybe<DataIndex>;
  dataIndices?: Maybe<Connection_DataIndex>;
  datasource?: Maybe<Datasource>;
  datasources?: Maybe<Connection_Datasource>;
  docType?: Maybe<DocType>;
  docTypeField?: Maybe<DocTypeField>;
  docTypeFieldNotInAnnotator?: Maybe<Connection_DocTypeField>;
  docTypeFields?: Maybe<Connection_DocTypeField>;
  docTypeFieldsByParent?: Maybe<Connection_DocTypeField>;
  docTypeFieldsFromDocType?: Maybe<Connection_DocTypeField>;
  docTypeFieldsFromDocTypeByParent?: Maybe<Connection_DocTypeField>;
  docTypeFieldsNotInSorting?: Maybe<Connection_DocTypeField>;
  docTypeFieldsNotInTokenTab?: Maybe<Connection_DocTypeField>;
  docTypeTemplate?: Maybe<DocTypeTemplate>;
  docTypeTemplates?: Maybe<Connection_DocTypeTemplate>;
  docTypes?: Maybe<Connection_DocType>;
  embeddingModel?: Maybe<EmbeddingModel>;
  embeddingModels?: Maybe<Connection_EmbeddingModel>;
  enabledBucket?: Maybe<Bucket>;
  enrichItem?: Maybe<EnrichItem>;
  enrichItems?: Maybe<Connection_EnrichItem>;
  enrichPipeline?: Maybe<EnrichPipeline>;
  enrichPipelines?: Maybe<Connection_EnrichPipeline>;
  /** Returns the list of events */
  event?: Maybe<Array<Maybe<Event>>>;
  /** Get event data */
  eventData?: Maybe<Scalars['String']>;
  /** Returns the list of available options for the event */
  eventOptions?: Maybe<Array<Maybe<EventOption>>>;
  language?: Maybe<Language>;
  languages?: Maybe<Connection_Language>;
  largeLanguageModel?: Maybe<LargeLanguageModel>;
  largeLanguageModels?: Maybe<Connection_LargeLanguageModel>;
  pluginDriver?: Maybe<PluginDriver>;
  pluginDrivers?: Maybe<Connection_PluginDriver>;
  pluginDriversPageFilter?: Maybe<Page_PluginDriver>;
  queryAnalyses?: Maybe<Connection_QueryAnalysis>;
  queryAnalysis?: Maybe<QueryAnalysis>;
  queryParserConfig?: Maybe<QueryParserConfig>;
  queryParserConfigs?: Maybe<Connection_QueryParserConfig>;
  ragConfiguration?: Maybe<RagConfiguration>;
  ragConfigurations?: Maybe<Connection_RagConfiguration>;
  rule?: Maybe<Rule>;
  rules?: Maybe<Connection_Rule>;
  scheduler?: Maybe<Scheduler>;
  schedulers?: Maybe<Connection_Scheduler>;
  searchConfig?: Maybe<SearchConfig>;
  searchConfigs?: Maybe<Connection_SearchConfig>;
  sorting?: Maybe<Sorting>;
  sortings?: Maybe<Connection_Sorting>;
  suggestionCategories?: Maybe<Connection_SuggestionCategory>;
  suggestionCategory?: Maybe<SuggestionCategory>;
  tab?: Maybe<Tab>;
  tabs?: Maybe<Connection_Tab>;
  tokenFilter?: Maybe<TokenFilter>;
  tokenFilters?: Maybe<Connection_TokenFilter>;
  tokenTab?: Maybe<TokenTab>;
  tokenTabs?: Maybe<Connection_TokenTab>;
  tokenizer?: Maybe<Tokenizer>;
  tokenizers?: Maybe<Connection_Tokenizer>;
  totalSortings?: Maybe<Connection_Sorting>;
  totalTokenTabs?: Maybe<Connection_TokenTab>;
  unboundAnalyzersByCharFilter?: Maybe<Array<Maybe<Analyzer>>>;
  unboundAnalyzersByTokenFilter?: Maybe<Array<Maybe<Analyzer>>>;
  unboundBucketsByDatasource?: Maybe<Array<Maybe<Bucket>>>;
  unboundBucketsBySuggestionCategory?: Maybe<Array<Maybe<Bucket>>>;
  unboundBucketsByTab?: Maybe<Array<Maybe<Bucket>>>;
  /** Fetches DocTypeFields unbound to the provided SuggestionCategory ID with FieldType KEYWORD or I18N. */
  unboundDocTypeFieldsBySuggestionCategory?: Maybe<Array<Maybe<DocTypeField>>>;
  unboundEnrichPipelines?: Maybe<Array<Maybe<EnrichPipeline>>>;
  /**
   * Retrieves a list of RAGConfiguration entities of the specified RAGType
   * that are not yet associated with the given Bucket.
   *
   * This query returns all RAGConfiguration instances that:
   * - Have the specified RAGType.
   * - Are not currently linked to the provided bucketId.
   *
   * Arguments:
   * - `bucketId` (ID!): The ID of the Bucket for which to retrieve unbound RAGConfiguration entities.
   * - `ragType` (RAGType!): The type of RAGConfiguration to filter by.
   *
   * Returns:
   * - A list of RAGConfiguration entities that match the criteria.
   *
   */
  unboundRAGConfigurationByBucket?: Maybe<Array<Maybe<RagConfiguration>>>;
  unboundTabsByTokenTab?: Maybe<Array<Maybe<Tab>>>;
};
// ...
export type QueryAnalysis = {
  __typename?: 'QueryAnalysis';
  annotators?: Maybe<Connection_Annotator>;
  /** ISO-8601 */
  createDate?: Maybe<Scalars['DateTime']>;
  description?: Maybe<Scalars['String']>;
  id?: Maybe<Scalars['ID']>;
  /** ISO-8601 */
  modifiedDate?: Maybe<Scalars['DateTime']>;
  name?: Maybe<Scalars['String']>;
  rules?: Maybe<Connection_Rule>;
  stopWords?: Maybe<Scalars['String']>;
  stopWordsList?: Maybe<Array<Maybe<Scalars['String']>>>;
};
// ...
export type Rule = {
  __typename?: 'Rule';
  /** ISO-8601 */
  createDate?: Maybe<Scalars['DateTime']>;
  description?: Maybe<Scalars['String']>;
  id?: Maybe<Scalars['ID']>;
  lhs?: Maybe<Scalars['String']>;
  /** ISO-8601 */
  modifiedDate?: Maybe<Scalars['DateTime']>;
  name?: Maybe<Scalars['String']>;
  rhs?: Maybe<Scalars['String']>;
};
// ...
export type QueryAnalysisAssociationsQuery = { __typename?: 'Query', queryAnalysis?: { __typename?: 'QueryAnalysis', id?: string | null, annotators?: { __typename?: 'DefaultConnection_Annotator', edges?: Array<{ __typename?: 'DefaultEdge_Annotator', node?: { __typename?: 'Annotator', id?: string | null, name?: string | null } | null } | null> | null } | null, rules?: { __typename?: 'DefaultConnection_Rule', edges?: Array<{ __typename?: 'DefaultEdge_Rule', node?: { __typename?: 'Rule', id?: string | null, name?: string | null } | null } | null> | null } | null } | null };